Work around missing pagination headers
The backend currently is not sending the expected pagination headers
necessary for the pagination UI in the Dependency List[0].

This change works around that. It:

 - fetches the full, un-paginated list of dependencies on load, to find
   out the total number of them (this information is sufficient to build
   the pagination UI);
 - persists/updates the current page number in the store, rather than
   based on the (missing) response headers.

This should be reverted once the fix[1] for the backend is merged,
though there shouldn't be any problem with them coexisting; the frontend
will then simply be making one unnecessary request.

[0]: https://gitlab.com/gitlab-org/gitlab-ee/issues/10075
[1]: https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/14115