Fix calculation of used extra CI minutes
When a Group/Personal namespace is using the Free plan on GL.com then
the assigned CI minutes is read from
`Gitlab::CurrentSettings.shared_runners_minutes` and not from
`Namespace#shared_runners_minutes`.

In the previous version we were using
`Namespace#shared_runners_minutes_limit` which was returning `nil`
instead of 2000 (The current default CI minutes configured globally for
free plans in GL.com). This bug was affecting Pipelines and was showing
wrong information on the Pipeline Quote info for the Group/Personal
namespace.

Now we're using `Namespace#actual_shared_runners_minutes_limit` which
returns the globally configured value if the current Namespace has not
been configured with CI minutes.