Show design boards at previous versions in GraphQL
The GraphQL designs query can take an new `atVersion` argument:

    designs(atVersion: "gid://gitlab/DesignManagement::Version/1") {
      edges {
        node {
          image
        }
      }
    }

A new finder can filter designs to only those that existed at the given
version.

This allows a design board to reflect previous states.

DesignTypes can access the `atVersion` argument passed to the parent
connection in order to include the version sha in the image URI it
presents.

https://gitlab.com/gitlab-org/gitlab-ee/issues/9491