Remove support for checking legacy security reports
The checking of the existence of the legacy security reports
was using a significant number of SQL queries in the merge request
widget. This commit removes support for them so that they will
no longer be accessible from the widget, but the data remains
intact.

Part of https://gitlab.com/gitlab-org/gitlab-ce/issues/63228