Remove `multiple_code_owner_rules` feature flag
Before we were creating a single `ApprovalRule` called "Code
Owner". But when we introduced `multiple_code_owner_rules` we were
creating a rule per entry in the CODEOWNERS file and displaying that.

Having the feature flag meant having to support the 2 datamodels. This
removes the feature flag and gets rid of the legacy code that was
there to support it's off-state.