Change 'Todo' to 'To Do'
Currently, we label items to be done as "Todo." This is grammatically
incorrect and (therefore) confusing—especially to our Spanish-speaking
users for whom "todo" has a specific and unrelated meaning.

We should use "To Do" and always use it as singular (not "To Dos").

Updates to wording in a few places per MR (ee) discussion

Updating locale/gitlab.pot

Updates to wording in a few places per MR (ee) discussion

Updating locale/gitlab.pot