Geo: Increase HTTP read timeout of proxy requests to 60 s
For large pushes over SSH, we observed that it took some time for the
secondary server to hear back from the primary server after the POST
request to the `/api/v4/geo/proxy_git_push/ssh/push` endpoint. The
client would give up on the request and fail the entire push.

To give more room for these pushes to go through, bump the timeout
from the original 10 seconds to 60 seconds.