Fix: duplicates during weight sort
Problem: when weight sorting is applied then the same issue can appear
again when a user goes the next page. It happens because issues might
have the same weight and database does not guarantee the same order
unless we explicitly request it.

Solution: this problem was solved some time ago it following MR:
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/4878/diffs

However, EE version overrides the fix from CE option and we need to
apply it again to resolve the problem.