Add an Upgrade button to User's billings page
It adds an upgrade button to the user billings
page that redirects to the customers portal
upgrade checkout page.

If the user is not logged in at the portal he/she
will be automatically redirected after the login
process.

If the user is at the latest tier, the button
won't be presented and it'll fallback to the
existing plans table.