Implement authetication for Conan Packages API
Add new Conan API endpoint `/users/authenticate` that when valid
personal access tokens is provided (as password for HTTP Basic Auth)
will respond with the same access token in the response body.

For subsequent requests the token will be provided by the Conan client
as bearer in the HTTP_AUTHORIZATION header.

The downside of this approach is that the access token will be stored
locally by the Conan client in a SQLite database.