Inject EE modules in a CE compatible fashion
All instances of injecting an EE specific module have been changed to
use the new methods for this: prepend_if_ee, extend_if_ee, and
include_if_ee. This allows these lines to be included in CE, even when
the modules to inject do not exist. This in turn allows us to backport
these lines to CE and keep them there, instead of having to strip them
out.