Reject API pushes that violate CODEOWNERS
Applying only to ProtectedBranches, this is a squashed commit of 40+
commits building out support for this feature. We begin the process here
of moving from applying CODEOWNER validation rules on a project level,
to enabling CODEOWNER validations to run more granularly, by moving this
attribute to the ProtectedBranches themselves. A latter project will
follow that enable managing this feature via the UI, but in this commit
we allow for manipulation of this value via the API. Finally, we extend
CODEOWNER validation checks to closed MRs as well, as the rules are
attached to the merge request with each new ref in the changeset, rather
than being stateful to the time the action is initiated; the rules that
applied at the moment of creation are the rules we apply in the future.