Return any available cluster from PrometheusAdapterService
Change the PrometheusAdapterService to return any available cluster
when deployment_platform is not provided, instead of only cluster with
environment scope *.
This is so that the custom metrics form works if any cluster is
available.
Also, Serverless::FunctionsFinder has a bug where the wrong cluster
will be used when a cluster with scope * and another cluster with any
other scope (say staging) are present at the same time. If
FunctionsFinder.invocation_metrics is called with 'staging' as parameter
it will query the cluster with scope * instead.