Use X_if_ee injection methods for left-over lines
In MR https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/14863 we
changed various prepend/include/extend lines to use the new X_if_ee
methods for injecting EE modules. Unfortunately, we forgot to take care
of a variety of lines due to the use of inaccurate patterns to find
these lines.

This commit takes care of moving over the remaining lines, except for
any prepend lines in the ee/ directory as the changes are not needed
there (since this directory is only available in EE).