Remove validation of MR level approval rules in MR
This removes `MergeRequest#validate_approval_rule_source` which
was responsible to check whether the MR level approval rules
matches the existing project level rules.

Added the validation in `ApprovalMergeRequestRule` to check if the
project of `approval_project_rule` matches the merge request's
project. This way we can ensure that the source rule should still
be in the same project.

This will make the logic simpler for the upcoming API to create
MR level approval rules.