Extract Dependency Scanning reports Vuex module
This is equivalent to another [MR][sastMR] which split out the SAST
reports to a new module, except this is for the Dependency Scanning
reports.

The newly-extract module is *intentionally* very similar to the SAST
module. This is to ease [future work][dryIssue] to apply the DRY
principle to these modules once they're all extracted.

See this [issue][issue], or its parent [epic][epic] for more background.

[dryIssue]: https://gitlab.com/gitlab-org/gitlab-ee/issues/13327
[sastMR]: https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/14299
[issue]: https://gitlab.com/gitlab-org/gitlab-ee/issues/11979
[epic]: https://gitlab.com/groups/gitlab-org/-/epics/1425