Show warnings in performance bar
Right now, the performance bar is unopinionated: it will tell you how
many SQL queries a page ran, but not if that's good or bad. It also
hides away AJAX requests behind a dropdown, which means people are less
likely to see those requests.

This adds a way of highlighting warnings generated by the backend in
both places. In a request itself, a metric will have a warning item if
the backend thinks it's too slow. And in the request selector, we
highlight requests that have warnings associated.