Add cop to check GraphQL field/argument description
- Fix typo
- Remove reference to outdated files
- Restrict directories to `app/graphql` and `ee/app/graphql` via
`.rubocop.yml`
- Correct all missing descriptions

Add a comment to disable to the cop for every missing GraphQL
description. When folks backfill the description to satisfy the cop,
they can remove the line.

Since we are now including arguments, and (probably) enums,
this is just called GraphQL Descriptions unless it proves
to be confusing.