Hide job-link in DL-failure-alert if no jobPath
If a job for the dependency-list generation fails, and the result
does not contain a path to the job the alert on the client-side
should not contain a button.

This commit adds a check to make sure the button within the alert
only shows if it receivs a 'jobPath' property.