Fix dependency-list link visibility in side-nav
A user, having the right permissions, should see a link to the
dependency-list in a project's sidebar navigation.

A previous change moved that link under a new top-level navigation
item, which is only rendered if the user has access to the
security-dashboard. This prevents some users from seeing the link to
the dependency-list, even if they have the right permissions.

This commit adds extra logic to correct the behaviour.