Consolidate logic for historical max user count
`License#historical_max` leverages `HistoricalData#max_historical_user_
count` for calculation.  It also defaults to using a date from from a
year prior to the expiration date up to the expiration date.
Adds `License#prior_historical_max` as a way to continue to easily get
the historical max before the given license which is used for
`check_users_limit`.