Update Dependency List documentation link
The Dependency List recently gained a dedicated documentation page[1].
This MR updates the "Learn more" link of the Dependency List view to
point to that new documentation.

As it was, the anchor on the previous link was broken, although the
linked page about Dependency Scanning does have an entry about the
Dependency List.

[1]: https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/31147