Checkbox select repository storages
Repository storages selection was rather brittle, as a few wrong clicks
in the small selection box could lead to a storage not being used, or
actually being used.

This change updates the styling a bit to conform to other styling, uses
i18n for strings in this section, and uses checkboxes to select storage
nodes.

Conflicts:
	locale/gitlab.pot