Avoid Devise "401 Unauthorized" responses
Instead, in GitLab, we want to redirect to the sign-in page. This
happens for most requests, but when Devise can't work out the request
format, or for certain other kinds of request, it will send a very poor
401 Unauthorized instead.

This is acceptable in certain cases (XHR or git clone requests, say),
but should not happen for user-visible requests - they should be
redirected to sign in, instead.