Fix existing Uploader specs
Our current specs base are in many cases, not testing uploads correctly.

There are many variations on how uploads behave, depending on which
Uploader and which model it is related to, but because our test cases
have mocked important pieces, it was never testing them.

This fixes that by removing mocks and making the required adjustments.