Prevent versioning Designs that are unmodified
Our `SaveDesignsService` would previously version and commit designs
files that were not modified since their last version.

These would appear as updates in the version history for the design.

Besides the version history becoming a bit inaccurate and silly looking,
this may create future problems for us. For example, because annotations
use the existing `DiffNote` classes, we get errors when producing the
discussion threads for discussions started on designs that have not
changed in their "diff". There may be other issues lurking if we allow
unmodified files to be committed.

https://gitlab.com/gitlab-org/gitlab-ee/issues/24080