Allow SmartCard authentication to use SAN extensions
This will allow to use a SubjectAlternateName (SAN) to identify the
user of the certificate, by matching a SAN entry from the it, that
contains at least the RFC822Name (email of the user) and an
uniformResourceIdentifier (URI).

Like:
`subjectAltName=email:some@email.com,URI:https://gitlab.email.com`

In this case the URI is the fully qualified name used for the GitLab
instance and the email is the identity of the user within the instance.