Fix setting of weight of a new issue in board list
When a new issue were created directly from the issue board list, users
were not able to set the weight because the board's store `updateWeight`
checked whether `sidebarInfoEndpoint` were available. Since it was not
set properly, the frontend never attempted to update it.

To fix this, we refactor the issue code by introducing a `refreshData`
method that can be used to update the issue data in one place.

Closes https://gitlab.com/gitlab-org/gitlab-ee/issues/5877