Store the secondary Geo node ID
At this point, we are the primary node, and the secondary has
redirected to us. We need to store the secondary Geo node ID
because Gitaly will make a request to the /internal/post_receive
API endpoint after we’ve responded OK to Workhorse.

Later, in /internal/post_receive, we will fetch the node ID and
generate a replication lag message.