Remove unused eslint-plugin-html
With a quick check it was easily discoverable that the plugin is
currently not used.

```
node_modules/.bin/eslint --print-config app/ | grep html
```

It has been added with 8050d835d40 but removed with eb336d7179c for vue
files and completely removed with 3e9bb4ad314.

As this plugin is unused and actively hindering an eslint upgrade, we
should remove it.