Simplify end-to-end test setup
Besides that, this commit tries to fix an issue that happens on
staging, where user fabrication via API doesn't work, and it may
be due to the usage of let(:user) since other tests using
User.fabricate_via_api! don't use it in a let declaration.

In this commit I'm moving the user fabrication to the before block
and exposing it as an instance variable @user, so that it can be
used in the test itself.

Also, I have removed unnecessary steps that makes the test slower.