CloudFormation template for creating EKS clusters
Creates the cluster itself, a group of worker nodes, and the
networking rules to allow them to communicate.

Assumes that a VPC, an IAM service role, subnets and
security group have already been created and will be
passed as parameters.

The nodes must wait for the cluster to finish creating
as they need to register with the cluster on launch,
which will fail if the cluster has yet to finish creating.