Temporary instrumention for internal id lock
This adds a temporary counter `gitlab_internal_id_for_update_lock`
tracking the number of `ROW SHARE` locks on internal id records.

The counter has an additional dimension `usage` which indicates
what kind of internal id lock scope was used. Also, `changed` indicates
whether or not the actual counter value changed with the attempted
update.

This is a preparation for
https://gitlab.com/gitlab-org/gitlab/issues/30515