Fix incorrect selection of custom templates
Previously if multiple project templates had similar names, it was
possible the wrong template were used. For example, if you had three
different templates with the names:

1. docker
2. docker-image
3. docker-image-and-app

Clicking on `docker` would cause the search to match all three items,
and `CreateFromTemplateService` would just pick the first one returned
from the database.

Instead of using names, this commit uses the project ID to avoid this
ambiguity.

The backend support for this feature went in GitLab 12.4 via
https://gitlab.com/gitlab-org/gitlab/merge_requests/18178. This change
updates the frontend to use this.

Closes https://gitlab.com/gitlab-org/gitlab/issues/9146