Disable scheduling of productivity analytics
This migration times out on both GitLab's staging and production
environments, even when an index is added for the columns used. As we
are nearing release day we have decided to turn this migration into a
noop for the time being.

The background migration is not removed as some jobs may have been
scheduled (especially in dev environments). Keeping this code allows
those jobs to finish, and allows us to reschedule it in the future if
needed.