Show 20 labels in dropdown instead of 5
When pressing ~ to show a list of labels, only 5 are shown. This can be
annoying for the user if there are more than 5 labels that start with
the same word but they can't see all matching labels in the autocomplete
list. This commit changes the list to show a maximum of 20 instead of 5.