Limit snippet search scope in GitLab.com
The scope we use when searching for snippets it's
quite wide. It searchs in personal snippets and also
project snippets (in authorized projects and also visible
when not authorized).

When the dataset is quite big, like in GitLab.com, the requests
times out because they take around 30s.

While we address and rearrange the snippet search to reduce
this scope for every type of instance, we're going
to narrow it just for GitLab.com.

Now, we're going to search only in the user personal and in
authorized project snippets.