Clear staged changes _before_ redirection
Any local changes in WebIDE should be cleared _before_ redirecting
to the new merge request page. Earlier, it was happening just after
redirection, which causes onbeforeunload event to trigger and warns the
user that his unsaved changes will be lost, when in fact there are no
unsaved changes at this point.

The test for this case was falsely passing. I have modified it to
ensure that it always fails in the absence of this change.