Add nav link for Instance Security Dashboard
This adds a nav link for the [Instance Security Dashboard MVC][1], which
is behind the `security_dashboard` feature flag.

This also refactors slightly how the visibility of the Operations and
Environments dashboard links is handled. One result of this is that they
are now consistently visible/hidden between the dashboards dropdown on
larger screens, and the "More" dropdown on smaller screens.

Finally, the partials have been renamed, since they're no longer
specific to operations.

[1]: https://gitlab.com/gitlab-org/gitlab/issues/6953