Increase timeouts for Gitaly RPCs
This change increases the timeout of the following RPCs in Sidekiq from
55 seconds to 6 hours:

* CreateFork
* CreateRepositoryFromUrl
* FetchSourceBranch

Prior to https://gitlab.com/gitlab-org/gitlab/merge_requests/16926,
`default_timeout` was unlimited for Sidekiq jobs, but now it is 55
seconds. However, the above RPCs can take a long time to run.

To avoid timing out prematurely, we increase these values to long timeouts
to ensure that they can complete.

Closes https://gitlab.com/gitlab-com/gl-infra/production/issues/1187