Update test with an additional expectation
I have noticed in the CI environment that sometimes this test
fails, but based on the taken screenshot, 2 participants are shown
in this section of the right sidebar, but only one assignee is
shown in the assignees list, which makes me think that it may be
taking some time for the avatar of the new assignee to show in
the assignees list.

The idea here is having an extra expectation before the one where
the count happens to give time for the second avatar to be shown
in the list of assignees.

I'll run the test a few times on CI before asking for a review, to
ensure that the test is not flaky, and if it continues with the
flakiness, I'll quarantine it for further investigation