Update artifact size authorization
Now that we have project and namespace level setting
for max_artifacts_size, we need to update the authorization
for it in some of the runner related API endpoints.

This also adds a new helper method `#closest_setting`
which fetches the closest non-nil value for a given setting name.
This will be useful for other settings like max_pages_size.