Extend a test to cover the MR resolve button fix
See: https://gitlab.com/gitlab-org/gitlab/issues/32157

Adds a couple of steps to an existing test to provide additional
cover for the fix in the MR above.

Updates the test to be more consistent with our standards.

Updates Resource::Repository::Commit add or update files, and
updates the one test that was using that resource.