Do not start mirroring via API when paused
If a mirroring job fails and retries 14 times, it'll be paused (or
marked as a hard fail).

We are no longer allowing to unpause/reset
this via API as it leads to mirroring jobs retrying even though
it's expected to fail.

It'll still be possible to be done via the web UI by forcing an
update.