Adds an E2E test of multi-project pipelines
Uses a group runner rather than 2 project runners, so it also adds the
ability to override the token used when registering a runner (the
default is to use the token for the project attribute of the runner).

Adds the ability to remove a runner via the API, which also removes
the runner service (docker container).

Extracts the CI badge element and methods to a component that is
shared with the pipeline page object and the job page object.

Fixes a bug in `click_element` that meant text input was ignored.