Enable Google API retries for uploads
As stated in
https://github.com/googleapis/google-api-ruby-client#errors--retries,
enabling retries is strongly encouraged but disabled by default. Large
uploads that may hit timeouts will mainly benefit from this.

Note that this retry mechanism only kicks in if the HTTP command
encounters a server error, rate limit error, or transmission error.

This will help improve the reliability of backups and anything else that
goes through google-api-client. fog-google appears to be the main user
of this.

Closes https://gitlab.com/gitlab-org/gitlab/issues/23452